Move react components into components folder fixes #1814
