+++ app/experimenter/static/js/index.js
@@ -4,7 +4,8 @@ import bsSelect from "bootstrap-select";
 import popper from "popper.js";
 import React from "react";
 import ReactDOM from "react-dom";
-import DesignForm from "design-form";
+
+import DesignForm from "experimenter/components/DesignForm";
 
 window.jQuery = jQuery;
 window.$ = jQuery;
