+++ app/experimenter/static/js/scripts/detail-base.js
