+++ app/experimenter/static/js/scripts/detail-review.js
