+++ app/experimenter/static/js/scripts/edit-variants.js
@@ -1,0 +1,7 @@
+// Initialize the formset plugin
+jQuery(function($) {
+  $("#formset").formset({
+    animateForms: true,
+    reorderMode: "dom"
+  });
+});
