+++ app/experimenter/static/js/scripts/experiment-list-select.js
