+++ app/experimenter/templates/experiments/detail_base.html
@@ -313,7 +313,7 @@
 {% endblock %}
 
 {% block extrascripts %}
-  <script src="{% static "js/detail-base.js" %}"></script>
+  <script src="{% static "js/scripts/detail-base.js" %}"></script>
 {% endblock %}
 
 {% block page_id %}page-detail-view{% endblock %}
