+++ app/experimenter/templates/experiments/detail_review.html
@@ -123,5 +123,5 @@
 
 {% block extrascripts %}
   {{ block.super }}
-  <script src="{% static "js/detail-review.js" %}"></script>
+  <script src="{% static "js/scripts/detail-review.js" %}"></script>
 {% endblock %}
