+++ app/experimenter/templates/experiments/edit_design.html
@@ -50,7 +50,7 @@
 {% block extrascripts %}
   <script src="{% static "js/jquery.formset.js" %}"></script>
 
-  <script src="{% static "js/edit-variants.js" %}"></script>
+  <script src="{% static "js/scripts/edit-variants.js" %}"></script>
 {% endblock %}
 
 {% block page_id %}page-edit-variants{% endblock %}
