+++ app/experimenter/templates/experiments/edit_objectives.html
@@ -38,5 +38,5 @@
 {% endblock %}
 
 {% block extrascripts %}
-    <script src="{% static "js/edit-objectives.js" %}"></script>
+    <script src="{% static "js/scripts/edit-objectives.js" %}"></script>
 {% endblock %}
