+++ app/experimenter/templates/experiments/edit_risks.html
@@ -81,7 +81,7 @@
 {% endblock %}
 
 {% block extrascripts %}
-<script src="{% static "js/edit-risks.js" %}"></script>
+<script src="{% static "js/scripts/edit-risks.js" %}"></script>
 
 
 {% endblock %}
