+++ app/experimenter/templates/experiments/edit_timeline_population.html
@@ -103,8 +103,8 @@
   {% endif %}
 {% endblock %}
 
-{% block extrascripts %}
-  <script src="{% static "js/edit-variants.js" %}"></script>
-{% endblock %}
+{% block page_id %}page-edit-timeline-and-population{% endblock %}
 
-{% block page_id %}page-edit-timeline-and-population{% endblock %} 
+{% block extrascripts %}
+  <script src="{% static "js/scripts/bootstrap-select.js" %}"></script>
+{% endblock %}
