+++ app/experimenter/templates/experiments/list.html
@@ -195,8 +195,8 @@
 {% endblock %}
 
 {% block extrascripts %}
-  <script src="{% static "js/experiment-date-filter.js" %}"></script>
-  <script src="{% static "js/experiment-list-select.js" %}"></script>
+  <script src="{% static "js/scripts/experiment-date-filter.js" %}"></script>
+  <script src="{% static "js/scripts/experiment-list-select.js" %}"></script>
 {% endblock %}
 
 {% block page_id %}page-list-view{% endblock %}
